backend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
sqlalchemy>=2.0.30
pydantic>=2.9.0
pydantic-settings>=2.4.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.12
python-dotenv>=1.0.1
alembic>=1.14.0
httpx>=0.27.0
resend>=0.6.0
psycopg2-binary>=2.9.9
